back/python/token_destroyer/requirements.txt
fastapi~=0.110.2
uvicorn~=0.29.0
python-dotenv~=1.0.1
asyncio~=3.4.3
pymongo~=4.6.3
pydantic~=2.6.4